+++ Proyecto-springboot/src/main/java/co/edu/poli/tutorship/model/SMS.java
@@ -1,0 +1,32 @@
+package co.edu.poli.tutorship.model;
+
+
+	public class SMS {
+		
+		private String phoneNumberTo;
+		
+		private String body;
+		
+		public SMS () {
+			
+		}
+
+		public String getPhoneNumberTo() {
+			return phoneNumberTo;
+		}
+
+		public void setPhoneNumberTo(String phoneNumberTo) {
+			this.phoneNumberTo = phoneNumberTo;
+		}
+
+		public String getBody() {
+			return body;
+		}
+
+		public void setBody(String body) {
+			this.body = body;
+		}
+
+	}
+
+
