+++ Proyecto-springboot/src/main/java/co/edu/poli/tutorship/services/SMSService.java
@@ -1,0 +1,13 @@
+package co.edu.poli.tutorship.services;
+
+
+	import com.twilio.rest.api.v2010.account.Message;
+
+	import co.edu.poli.tutorship.model.SMS;
+
+	public interface SMSService {
+		
+		public Message sendSMS(SMS sms);
+
+	}
+
