+++ Proyecto-springboot/src/main/resources/application.properties
@@ -1,0 +1,37 @@
+# SSL (Secure Sockets Layer) and its successor, TLS (Transport Layer Security), 
+# are protocols for establishing authenticated and encrypted links between networked computers.
+spring.datasource.url = jdbc:mysql://localhost:3306/tutorship?useSSL=false
+spring.datasource.username=root
+spring.datasource.password=root
+
+spring.mail.host=smtp.office365.com
+spring.mail.port=587
+spring.mail.username=paperezt@poligran.edu.co
+spring.mail.password=kiarayani1234567
+spring.mail.properties.mail.smpt.auth=true
+spring.mail.properties.mail.smtp.connecttimeout=5000
+spring.mail.properties.mail.smtp.timeout=3000
+spring.mail.properties.mail.smtp.writetimeout=5000
+spring.mail.properties.mail.smtp.starttls.enable=true
+# The dialect specifies the type of database used in hibernate so that hibernate generate 
+# appropriate type of SQL statements.
+# ORACLE
+# org.hibernate.dialect.OracleDialect
+# MYSQL
+# org.hibernate.dialect.MySQLDialect
+# org.hibernate.dialect.MySQLInnoDBDialect  (ACID: Atomicity, Consistency, Isolation, and Durability)
+# org.hibernate.dialect.MySQLMyISAMDialect
+# SQL SERVER
+# org.hibernate.dialect.SQLServerDialect
+# POSTGRESS
+# org.hibernate.dialect.ProgressDialect
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
+
+# create :- create the schema, the data previously present (if there) in the schema is lost
+# update:- update the schema with the given values.
+# validate:- validate the schema. It makes no change in the DB.
+# create-drop:- create the schema with destroying the data previously present(if there). 
+#                It also drop the database schema when the SessionFactory is closed.
+spring.jpa.hibernate.ddl-auto=update
+
+spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
